Return definitions data as frozendict/frozenlist where possible [RHELDST-5671]
CDN Definitions should be immutable for anyone consuming them, to prevent unintentional
(or even intentional, but wildly incorrect) modification on-the-fly while consuming the
definitions. This commit tries to ensure the immutability by returning all dictionaries
as frozendicts and all lists an frozenlists (using frozenlist2 module)

This change is compatible even with python 2.6, however el6eng and el7eng packages
of frozendict and frozenlist2 don't exist and there is no intent to create them, so
even though the requirements for these are are listed in requirements.txt without
restrictions, there is a fallback to builtin dict and/or list if they are not installed.